+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 ## Version 4.0.1
 
 - fixed a bug where casts of friendly players would show up when using Friendly Nameplates
+- fixed a bug where any alpha color change was not persisted